Moved util file
